Begin using PullPolicy from config
