+++ internal/commands/build_test.go
@@ -152,6 +152,33 @@ func testBuildCommand(t *testing.T, when spec.G, it spec.S) {
 			})
 		})
 
+		when("--pull-policy is not provided", func() {
+			when("no pull policy set in config", func() {
+				it("uses the default policy", func() {
+					mockClient.EXPECT().
+						Build(gomock.Any(), EqBuildOptionsWithPullPolicy(pubcfg.PullAlways)).
+						Return(nil)
+
+					command.SetArgs([]string{"image", "--builder", "my-builder"})
+					h.AssertNil(t, command.Execute())
+				})
+			})
+			when("pull policy is set in config", func() {
+				it("uses the set policy", func() {
+					mockClient.EXPECT().
+						Build(gomock.Any(), EqBuildOptionsWithPullPolicy(pubcfg.PullNever)).
+						Return(nil)
+
+					cfg := config.Config{PullPolicy: "never"}
+					command := commands.Build(logger, cfg, mockClient)
+
+					logger.WantVerbose(true)
+					command.SetArgs([]string{"image", "--builder", "my-builder"})
+					h.AssertNil(t, command.Execute())
+				})
+			})
+		})
+
 		when("volume mounts are specified", func() {
 			it("mounts the volumes", func() {
 				mockClient.EXPECT().
